requirements.txt
click==8.0.4
cv==1.0.0
cycler==0.11.0
decorator==4.4.2
dill==0.3.4
joblib==1.1.0
kiwisolver==1.3.1
llvmlite==0.38.0
matplotlib==3.3.4
multiprocess==0.70.12.2
networkx==2.5.1
numba==0.55.1
numpy==1.19.5
pandas==1.4.2
pathos==0.2.8
Pillow==8.4.0
pox==0.3.0
ppft==1.6.6.4
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
tifffile==2020.9.3
toml==0.10.2
tqdm==4.63.0
typing_extensions==4.1.1
zipp==3.6.0
